quaternion float template instance

+++ core/vnl/Templates/vnl_quaternion+float-.cxx
@@ -1,4 +1,4 @@
 
 #include <vnl/vnl_quaternion.txx>
 
-VNL_QUATERNION_INSTANTIATE(double);
+VNL_QUATERNION_INSTANTIATE(float);
